REFACTOR: Modified method for train-test splitting
Renamed function from `split_lists` to `split_list`
because I do not foresee using multiple lists of
facilities. If we need to do this, it is easy to
add them all together before splitting.

If there is only 1 item in the input list, then
we are returning that item for both the training
and the testing data.

Ensuring that there is always at least 1 testing
item returned, even if only two items are passed.

Finally, made the train-test split percentage an
input parameter for easy user access.